#490645: used tiny int

+++ database/migrations/2025_05_09_061002_add_is_package_lesson_if_not_exists_to_lessons_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         if (!Schema::hasColumn('lessons', 'is_package_lesson')) {
             Schema::table('lessons', function (Blueprint $table) {
-                $table->smallInteger('is_package_lesson')->default(1);
+                $table->tinyInteger('is_package_lesson')->default(1);
             });
         }
     }
